# Conflicts:
#	core/src/common.ts
#	core/src/evaluation/common.ts
#	core/src/evaluation/conversation_scenarios.ts
#	core/src/evaluation/simulation/pre_built_personas.ts
#	core/src/evaluation/simulation/user_simulator_personas.ts
#	core/test/evaluation/conversation_scenarios_own_test.ts
#	core/test/evaluation/conversation_scenarios_test.ts
#	core/test/evaluation/simulation/pre_built_personas_own_test.ts
#	core/test/evaluation/simulation/pre_built_personas_test.ts
#	core/test/evaluation/simulation/user_simulator_personas_own_test.ts
#	core/test/evaluation/simulation/user_simulator_personas_test.ts
#	docs/guides/README.md
#	docs/guides/evaluation/conversation_scenarios/index.md

+++ core/src/evaluation/conversation_scenarios.ts
@@ -120,7 +120,7 @@ export const conversationScenariosModel: EvalModel<ConversationScenarios> =
 export const conversationGenerationConfigModel: EvalModel<ConversationGenerationConfig> =
   evalModel(
     {
-      count: z.number(),
+      count: z.number().int(),
       generationInstruction: optionalField(z.string()),
       environmentContext: optionalField(z.string()),
       modelName: z.string(),
